Move graalvm into java-simple-application generator
