+++ generators/java-simple-application/application.ts
@@ -29,9 +29,9 @@ import {
   SERVER_TEST_RES_DIR,
   SERVER_TEST_SRC_DIR,
 } from '../generator-constants.js';
-import { GRAALVM_REACHABILITY_METADATA } from '../java/generators/graalvm/internal/constants.ts';
 import { getMainClassName } from '../java/support/util.ts';
 
+import { GRAALVM_REACHABILITY_METADATA } from './generators/graalvm/internal/constants.ts';
 import type { Application } from './types.ts';
 
 export type JavaSimpleApplicationAddedApplicationProperties = {
