+++ generators/java-simple-application/generators/graalvm/command.ts
