+++ generators/java-simple-application/generators/graalvm/generator.spec.ts
