+++ generators/java-simple-application/generators/graalvm/index.ts
