+++ generators/java-simple-application/generators/graalvm/internal/constants.ts
