+++ generators/java-simple-application/generators/graalvm/internal/maven-definition.ts
