+++ generators/java-simple-application/generators/graalvm/resources/build.gradle
