+++ generators/java-simple-application/generators/graalvm/resources/gradle/libs.versions.toml
