+++ generators/java-simple-application/generators/graalvm/templates/README.md.jhi.native.ejs
