+++ generators/java-simple-application/generators/graalvm/templates/gradle/native.gradle.ejs
