+++ generators/java-simple-application/generators/graalvm/templates/src/main/java/_package_/config/NativeConfiguration.java.ejs
