+++ lib/index.spec.ts
@@ -72,11 +72,11 @@ describe('getGeneratorNamespaces', () => {
   "java-simple-application",
   "java-simple-application:bootstrap",
   "java-simple-application:code-quality",
+  "java-simple-application:graalvm",
   "java-simple-application:jib",
   "java:bootstrap",
   "java:build-tool",
   "java:domain",
-  "java:graalvm",
   "java:i18n",
   "java:node",
   "java:openapi-generator",
