+++ renovate.json
@@ -14,7 +14,7 @@
     {
       "customType": "regex",
       "description": "Update constants in js/ts files",
-      "fileMatch": ["^generators/java/generators/graalvm/internal/constants.ts$"],
+      "fileMatch": ["^generators/java-simple-application/generators/graalvm/internal/constants.ts$"],
       "matchStrings": [
         "// renovate: datasource=(?<datasource>[a-z-]+?)(?: depName=(?<depName>.+?))? packageName=(?<packageName>.+?)(?: versioning=(?<versioning>[a-z-]+?))?\\sexport const (?:.+) = '(?<currentValue>[\\w+\\.\\-]*)'"
       ]
